Simplify `didFocus` logic in `TextEditorComponent`
Prior to #15302, we wanted to make sure the component had rendered the
hidden input when receiving the focus event. To do so, we added some
workarounds for scenarios where the focus event was triggered before the
component had the chance to detect it was attached or visible.

After that pull-request, however, we always render the hidden input
independently of which events the component has observed, thus making
those workarounds not necessary anymore.

Please, note that we decided not to include this commit's changes in
beta, and want to let them bake for a full release cycle instead.